Safari-609.1.11.1/LayoutTests/media/track/in-band/track-in-band-mkv-chapters-expected.txt


Check for chapter track from an mkv file.

EVENT(canplaythrough)

** Setting track 1 to showing
RUN(inbandTrack1.mode = 'showing')
RUN(video.play())
EXPECTED (inbandTrack1.cues.length > '0') OK
RUN(video.pause())
RUN(video.currentTime = 0)
RUN(video.play())

EVENT(seeked)
EXPECTED (inbandTrack1.cues.length > '0') OK
RUN(video.pause())
Test all cues are equal OK
END OF TEST
